Updated WikiParser.java
It is now working with the maven framework.  In addition, the name of the files to be read is now passed as an argument into the main function.  This may require some changes in the future depending on whether we want to read multiple files at once but it will be a quick fix.