Add files via upload
The Serialize class is implemented to provide helper functions to serialize list of WikiPage objects as well as to deserialize required files.  This  may require future changes depending on our requirements.  The WikiExtractParser class allows the wiki text files in a specified folder to be parsed in a multi-threaded way.